+++ flowman-kernel-server/src/main/scala/com/dimajix/flowman/kernel/service/MultiWorkspaceManager.scala
@@ -16,10 +16,14 @@
 
 package com.dimajix.flowman.kernel.service
 
+import java.util.UUID
+
 import scala.collection.mutable
+import scala.util.control.NonFatal
 
 import org.slf4j.LoggerFactory
 
+import com.dimajix.common.ExceptionUtils.reasons
 import com.dimajix.flowman.fs.File
 import com.dimajix.flowman.spec.storage.SimpleWorkspace
 import com.dimajix.flowman.storage.Workspace
@@ -27,61 +31,87 @@ import com.dimajix.flowman.storage.Workspace
 
 class MultiWorkspaceManager(root:File) extends WorkspaceManager {
     private val logger = LoggerFactory.getLogger(classOf[MultiWorkspaceManager])
-    private val workspaces: mutable.Map[String, SimpleWorkspace] = mutable.Map()
+    private val workspaces: mutable.Map[String, (Option[UUID],SimpleWorkspace)] = mutable.Map()
 
     logger.info(s"Initialized MultiWorkspaceManager at '$root'")
 
     // Create workspace are + default workspace
     root.mkdirs()
-    workspaces.put("default", SimpleWorkspace.create(root / "default"))
+    workspaces.put("default", (None, SimpleWorkspace.create(root / "default")))
 
 
-    override def list(): Seq[Workspace] = {
-        val ws = SimpleWorkspace.list(root)
-
+    override def list(): Seq[(String,Workspace)] = {
+        val allWorkspaces = SimpleWorkspace.list(root)
+        val allNames = allWorkspaces.map(_.name).toSet
         workspaces.synchronized {
-            workspaces.clear()
-            ws.foreach(ws => workspaces.put(ws.name, ws))
+            allWorkspaces.foreach { ws =>
+                if (!workspaces.contains(ws.name)) {
+                    workspaces.put(ws.name, (None, ws))
+                }
+            }
+
+            val oldNames = workspaces.keySet -- allNames
+            oldNames.foreach(workspaces.remove)
         }
 
-        ws
+        workspaces.map { case (id,ws) => id -> ws._2 }.toSeq
     }
 
-    override def createWorkspace(name: String): Workspace = {
-        val path = root / name
-        if (SimpleWorkspace.exists(path))
-            throw new IllegalArgumentException(s"Workspace '$name' already exists")
+    override def createWorkspace(id: String, name:String, clientId:Option[UUID]): (String,Workspace) = {
+        val path = root / id
+        if (SimpleWorkspace.exists(path)) {
+            throw new IllegalArgumentException(s"Workspace '$id' already exists")
+        }
+        logger.info(s"Creating workspace '$id' for client '${clientId.getOrElse("")}'")
         val ws = SimpleWorkspace.create(path)
 
         workspaces.synchronized {
-            workspaces.put(name, ws)
+            workspaces.put(id, (clientId, ws))
         }
 
-        ws
+        id -> ws
     }
 
-    override def deleteWorkspace(name: String): Unit = {
-        val path = root / name
+    override def deleteWorkspace(id: String): Unit = {
+        val path = root / id
         if (!SimpleWorkspace.exists(path))
-            throw new IllegalArgumentException(s"Workspace '$name' does not exists")
+            throw new IllegalArgumentException(s"Workspace '$id' does not exists")
 
-        logger.info(s"Deleting workspace '$name' at '$path'")
+        logger.info(s"Deleting workspace '$id' at '$path'")
         workspaces.synchronized {
-            workspaces.remove(name)
+            workspaces.remove(id)
         }
 
         path.delete(true)
     }
 
-    override def getWorkspace(name: String): Workspace = {
-        workspaces.synchronized {
-            workspaces.getOrElseUpdate(name, {
-                val path = root / name
+    override def getWorkspace(id: String): (String,Workspace) = {
+        val ws = workspaces.synchronized {
+            workspaces.getOrElseUpdate(id, {
+                val path = root / id
                 if (!SimpleWorkspace.exists(path))
-                    throw new IllegalArgumentException(s"Workspace '$name' does not exists")
+                    throw new IllegalArgumentException(s"Workspace '$id' does not exists")
                 val ws = SimpleWorkspace.load(path)
-                ws
+                None -> ws
             })
+        }
+
+        id -> ws._2
+    }
+
+    override def removeClientWorkspaces(clientId: UUID): Unit = {
+        val toBeRemoved = workspaces.synchronized {
+            workspaces.filter(_._2._1.contains(clientId)).keys.toSeq
+        }
+
+        toBeRemoved.foreach { ws =>
+            try {
+                deleteWorkspace(ws)
+            }
+            catch {
+                case NonFatal(ex) =>
+                    logger.warn(s"Error removing client specific workspace '$ws':\n  ${reasons(ex)}")
+            }
         }
     }
 }