+++ CHANGELOG.md
@@ -60,6 +60,7 @@ Breaking changes will be documented in this changelog file for each version.
 * github-311: Create integration test for EMR
 * github-362: Upgrade EMR to 6.10
 * github-369: Prevent endless loop in Kernel client, when getContext fails
+* github-370: Kernel client should use temporary workspaces with automatic cleanup
 
 ### Breaking changes